requirements.txt
flask
flask_cors
requests
bs4
reportlab
pymongo
snowflake-connector-python
boto3